requirements.txt
beautifulsoup4==4.12.2
IMDbPY==2022.7.9
pandas==2.0.2
python-dateutil==2.8.2
pywin32==306
requests==2.31.0
requests-oauthlib==1.3.1
urllib3==2.0.3